[src] Cosmetic edit
